+++ src/error.rs
@@ -1,44 +1,43 @@
 use std;
+use std::borrow::Cow;
 
+/// `ErrorKind` specifies a high-level error category.
+#[derive(Clone, Copy, Debug, Eq, Hash, Ord, PartialEq, PartialOrd)]
+pub enum ErrorKind {
+    /// The input is invalid.
+    InvalidInput,
+
+    #[doc(hidden)]
+    Unspecified,
+}
+
+/// `Error` stores all information pertaining to an error.
 #[derive(Debug)]
 pub struct Error {
     kind: ErrorKind,
-    payload: ErrorPayload,
-}
-
-#[derive(Debug)]
-enum ErrorPayload {
-    Recursive(Box<std::error::Error>),
-    DynamicMessage(String),
-    StaticMessage(&'static str),
-}
-
-#[derive(Clone, Copy, Debug, Eq, Hash, PartialEq)]
-pub enum ErrorKind {
-    BadName,
-    #[doc(hidden)]
-    Other,
+    reason: Option<Cow<'static, str>>,
+    cause: Option<Box<std::error::Error>>,
 }
 
 impl Error {
-    pub fn new_with_dynamic_message<M: Into<String>>(kind: ErrorKind, message: M) -> Self {
+    pub fn new<R>(kind: ErrorKind, reason: R) -> Self
+        where R: Into<Cow<'static, str>>
+    {
         Error {
             kind: kind,
-            payload: ErrorPayload::DynamicMessage(message.into()),
+            reason: Some(reason.into()),
+            cause: None,
         }
     }
 
-    pub fn new_with_static_message(kind: ErrorKind, message: &'static str) -> Self {
+    pub fn new_with_cause<E, R>(kind: ErrorKind, reason: R, cause: E) -> Self
+        where E: Into<Box<std::error::Error>>,
+              R: Into<Cow<'static, str>>
+    {
         Error {
             kind: kind,
-            payload: ErrorPayload::StaticMessage(message),
-        }
-    }
-
-    pub fn new_with_inner_error<E: Into<Box<std::error::Error>>>(kind: ErrorKind, inner_error: E) -> Self {
-        Error {
-            kind: kind,
-            payload: ErrorPayload::Recursive(inner_error.into()),
+            reason: Some(reason.into()),
+            cause: Some(cause.into()),
         }
     }
 
@@ -49,34 +48,40 @@ impl Error {
 
 impl std::fmt::Display for Error {
     fn fmt(&self, f: &mut std::fmt::Formatter) -> Result<(), std::fmt::Error> {
-        match self.payload {
-            ErrorPayload::Recursive(ref e) => e.fmt(f),
-            ErrorPayload::DynamicMessage(ref m) => m.fmt(f),
-            ErrorPayload::StaticMessage(m) => m.fmt(f),
+        match (&self.reason, &self.cause) {
+            (&Some(ref reason), &Some(ref cause)) => write!(f, "{}: {}", reason, cause),
+            (&Some(ref reason), &None) => reason.fmt(f),
+            (&None, &Some(ref cause)) => cause.fmt(f),
+            (&None, &None) => unreachable!(),
         }
     }
 }
 
 impl std::error::Error for Error {
     fn description(&self) -> &str {
-        match self.payload {
-            ErrorPayload::Recursive(ref e) => e.description(),
-            ErrorPayload::DynamicMessage(ref m) => &m,
-            ErrorPayload::StaticMessage(m) => m,
+        use std::ops::Deref;
+        match (&self.reason, &self.cause) {
+            (&Some(ref reason), _) => reason.deref(),
+            (_, &Some(ref cause)) => cause.description(),
+            (&None, &None) => unreachable!(),
         }
     }
 
     fn cause(&self) -> Option<&std::error::Error> {
         use std::ops::Deref;
-        match self.payload {
-            ErrorPayload::Recursive(ref e) => Some(e.deref()),
-            _ => None,
+        match self.cause {
+            Some(ref e) => Some(e.deref()),
+            None => None,
         }
     }
 }
 
 impl From<std::io::Error> for Error {
     fn from(e: std::io::Error) -> Self {
-        Error::new_with_inner_error(ErrorKind::Other, e)
+        Error {
+            kind: ErrorKind::Unspecified,
+            reason: None,
+            cause: Some(Box::new(e)),
+        }
     }
 }